feat(creator): segment smartassets / messages / events
