+++ apps/angular-creator/src/app/components/actions/action-create-and-store-message/action-create-and-store-message.component.ts
@@ -46,7 +46,7 @@ export class ActionCreateAndStoreMessageComponent implements Action {
 
     try {
       this.loading = true;
-      const createdMessage = await this.creator.createAndStoreMessage({
+      const createdMessage = await this.creator.messages.createAndStoreMessage({
         smartAssetId: parseInt(this.smartAssetId),
         content,
         messageId: this.id && this.id !== '' ? parseInt(this.id) : undefined,
